requirements.txt
alabaster==0.7.12
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
Babel==2.8.0
beautifulsoup4==4.6.0
boto3==1.12.26
botocore==1.15.26
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
CommonMark==0.5.4
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-heroku==0.3.1
django-modelcluster==5.0.1
django-rest-framework==0.1.0
django-storages==1.8
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
docutils==0.15.2
draftjs-exporter==2.1.7
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.17.0
idna==2.9
imagesize==1.2.0
isort==4.3.21
Jinja2==2.11.1
jmespath==0.9.5
l18n==2018.5
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==4.1.3
Pillow==6.2.2
protobuf==3.11.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
recommonmark==0.4.0
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
snowballstemmer==2.0.0
Sphinx==1.4.9
sphinx-rtd-theme==0.4.3
sqlparse==0.3.0
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
wrapt==1.11.2
